fix(ci): clear the remaining matrix failures found in the first real CI run
Six distinct causes, each traced to a job log:

- tests/src/core/PhaseDMixerTestSupport.h: evidence() passes a non-literal
  format to fprintf; -Wformat-security (an error under -DUSE_WERROR) broke
  every Linux job. Suppressed locally in the helper.
- include/RecordRingBuffer.h: std::min(size_t, uint64_t) — m_capacity is
  size_t while the ring positions are uint64_t. Identical on Linux (both
  unsigned long), a deduction conflict on macOS (unsigned long long).
  Explicit cast, matching the cast already used at line 124.
- tests/src/plugins/PluginPortsMigrationTest.cpp: the slice-8 check asserted a
  plugin was present; a runner without libgig/STK/lilv/fluidsynth builds none,
  failing the whole suite. QSKIP instead of QVERIFY.
- plugins/Xpressive/CMakeLists.txt: /bigobj was only applied on the Windows
  targets, so msvc-x64 died with C1128 (exprtk past the section limit).
- tests/CMakeLists.txt: the partc_ref_* harness targets link only Qt, while
  every real plugin links the lmms target (BuildPlugin.cmake:54). That is
  invisible on Unix but needs the import library on Windows → mingw64 and
  windows-arm64 failed with undefined lmms::gui::* / embed::getIconPixmap.
  Also mark the libgig includes SYSTEM in the reference target, mirroring
  plugins/GigPlayer/CMakeLists.txt (libgig's DLS.h trips
  -Woverloaded-virtual, which -Werror promotes even from a system header).
- src/gui/MainWindow.cpp: two direct QMenu::addAction call sites still used the
  overload Qt 6.3 deprecated (-Wdeprecated-declarations, an error under
  -DUSE_WERROR with Qt6); upstream guards this elsewhere in the same file.

Verified locally as far as the toolchain allows: configure 0; the reference
target link is validated by building partc_ref_amplifier; the CI-only causes
(CMP0177, MSVC, mingw) cannot be reproduced on this box — CI is the verifier.

+++ include/RecordRingBuffer.h
@@ -100,7 +100,7 @@ public:
 	{
 		const auto writePos = m_writePos.load(std::memory_order_relaxed);
 		const auto readPos = m_readPos.load(std::memory_order_acquire);
-		const auto freeFrames = m_capacity - (writePos - readPos);
+		const auto freeFrames = m_capacity - static_cast<std::size_t>(writePos - readPos);
 		const auto toWrite = std::min(frames, freeFrames);
 		for (std::size_t i = 0; i < toWrite; ++i)
 		{
